Refactor tests, so unit tests are in their modules #31
