+++ src/octree.rs
@@ -828,4 +828,152 @@ mod tests {
         assert_eq!(results.len(), 1);
         assert_eq!(results[0], target);
     }
+
+    #[test]
+    fn test_delete_existing_point() {
+        let boundary = Cube {
+            x: 0.0,
+            y: 0.0,
+            z: 0.0,
+            width: 100.0,
+            height: 100.0,
+            depth: 100.0,
+        };
+        let mut tree: Octree<&str> = Octree::new(&boundary, 2).unwrap();
+        let p1 = Point3D::new(10.0, 10.0, 10.0, Some("A"));
+        let p2 = Point3D::new(20.0, 20.0, 20.0, Some("B"));
+        tree.insert(p1.clone());
+        tree.insert(p2);
+
+        assert!(tree.delete(&p1));
+        let results = tree.knn_search::<EuclideanDistance>(&p1, 1);
+        assert_ne!(results[0], p1);
+        assert!(!tree.delete(&p1));
+    }
+
+    #[test]
+    fn test_empty_tree_queries() {
+        let boundary = Cube {
+            x: 0.0,
+            y: 0.0,
+            z: 0.0,
+            width: 100.0,
+            height: 100.0,
+            depth: 100.0,
+        };
+        let mut tree: Octree<&str> = Octree::new(&boundary, 2).unwrap();
+        let target = Point3D::new(5.0, 5.0, 5.0, None::<&str>);
+
+        let knn_results = tree.knn_search::<EuclideanDistance>(&target, 5);
+        assert!(knn_results.is_empty());
+
+        let range_results = tree.range_search::<EuclideanDistance>(&target, 10.0);
+        assert!(range_results.is_empty());
+
+        assert!(!tree.delete(&target));
+    }
+
+    #[test]
+    fn test_knn_edge_cases() {
+        let boundary = Cube {
+            x: 0.0,
+            y: 0.0,
+            z: 0.0,
+            width: 100.0,
+            height: 100.0,
+            depth: 100.0,
+        };
+        let mut tree: Octree<&str> = Octree::new(&boundary, 4).unwrap();
+        let points = vec![
+            Point3D::new(10.0, 10.0, 10.0, Some("A")),
+            Point3D::new(20.0, 20.0, 20.0, Some("B")),
+            Point3D::new(30.0, 30.0, 30.0, Some("C")),
+        ];
+        let num_points = points.len();
+        tree.insert_bulk(&points);
+
+        let target = Point3D::new(15.0, 15.0, 15.0, None::<&str>);
+        let knn_results = tree.knn_search::<EuclideanDistance>(&target, 0);
+        assert!(knn_results.is_empty());
+
+        let knn_results = tree.knn_search::<EuclideanDistance>(&target, num_points + 5);
+        assert_eq!(knn_results.len(), num_points);
+    }
+
+    #[test]
+    fn test_duplicates_delete_one() {
+        let boundary = Cube {
+            x: 0.0,
+            y: 0.0,
+            z: 0.0,
+            width: 100.0,
+            height: 100.0,
+            depth: 100.0,
+        };
+        let mut tree: Octree<&str> = Octree::new(&boundary, 4).unwrap();
+        let p1 = Point3D::new(10.0, 10.0, 10.0, Some("A"));
+        let p2 = p1.clone();
+        tree.insert(p1.clone());
+        tree.insert(p2.clone());
+
+        let results = tree.knn_search::<EuclideanDistance>(&p1, 2);
+        assert_eq!(results.len(), 2);
+
+        assert!(tree.delete(&p1));
+        let results_after_delete = tree.knn_search::<EuclideanDistance>(&p1, 2);
+        assert_eq!(results_after_delete.len(), 1);
+    }
+
+    #[test]
+    fn test_range_search_includes_boundary_point() {
+        let boundary = Cube {
+            x: 0.0,
+            y: 0.0,
+            z: 0.0,
+            width: 100.0,
+            height: 100.0,
+            depth: 100.0,
+        };
+        let mut tree: Octree<&str> = Octree::new(&boundary, 4).unwrap();
+        let center = Point3D::new(50.0, 50.0, 50.0, Some("C"));
+        let boundary_point = Point3D::new(60.0, 50.0, 50.0, Some("B"));
+        tree.insert(center.clone());
+        tree.insert(boundary_point.clone());
+
+        let results = tree.range_search::<EuclideanDistance>(&center, 10.0);
+        assert!(results.contains(&boundary_point));
+        assert!(results.contains(&center));
+    }
+
+    #[test]
+    fn test_bulk_insert_empty_noop() {
+        let boundary = Cube {
+            x: 0.0,
+            y: 0.0,
+            z: 0.0,
+            width: 100.0,
+            height: 100.0,
+            depth: 100.0,
+        };
+        let mut tree: Octree<i32> = Octree::new(&boundary, 4).unwrap();
+        let empty: Vec<Point3D<i32>> = Vec::new();
+        tree.insert_bulk(&empty);
+        let target = Point3D::new(10.0, 10.0, 10.0, None::<i32>);
+        let results = tree.knn_search::<EuclideanDistance>(&target, 1);
+        assert!(results.is_empty());
+    }
+
+    #[test]
+    fn test_zero_capacity_rejected() {
+        let boundary = Cube {
+            x: 0.0,
+            y: 0.0,
+            z: 0.0,
+            width: 100.0,
+            height: 100.0,
+            depth: 100.0,
+        };
+        let result = Octree::<i32>::new(&boundary, 0);
+        assert!(result.is_err());
+    }
 }
