+++ tests/test_integration_rstar_tree.rs
@@ -227,169 +227,6 @@ fn test_rstar_tree_insert_bulk_2d() {
 }
 
 #[test]
-fn test_rstar_tree_forced_reinsertion() {
-    let mut tree: RStarTree<Point2D<i32>> = RStarTree::new(4).unwrap();
-    let points: Vec<_> = (0..5)
-        .map(|i| Point2D::new(i as f64, i as f64, Some(i)))
-        .collect();
-
-    for p in &points {
-        tree.insert(p.clone());
-    }
-
-    assert_eq!(
-        tree.height(),
-        2,
-        "Tree height should be 2 after 5 insertions"
-    );
-
-    // Insert more points to trigger overflow in a child node.
-    for i in 5..10 {
-        tree.insert(Point2D::new(i as f64, i as f64, Some(i)));
-    }
-
-    assert_eq!(
-        tree.height(),
-        2,
-        "Tree height should still be 2 after forced reinsertion"
-    );
-
-    let all_points = tree.range_search_bbox(&spart::geometry::Rectangle {
-        x: -1.0,
-        y: -1.0,
-        width: 11.0,
-        height: 11.0,
-    });
-    assert_eq!(all_points.len(), 10, "All points should be in the tree");
-}
-
-#[test]
-fn test_rstar_tree_delete_underflow() {
-    let mut tree: RStarTree<Point2D<i32>> = RStarTree::new(4).unwrap();
-    let points: Vec<_> = (0..10)
-        .map(|i| Point2D::new(i as f64, i as f64, Some(i)))
-        .collect();
-
-    for p in &points {
-        tree.insert(p.clone());
-    }
-
-    // min_entries is (4 * 0.4).ceil() = 2.
-    // Deleting points to trigger underflow.
-    assert!(tree.delete(&points[0]));
-    assert!(tree.delete(&points[1]));
-    assert!(tree.delete(&points[2]));
-
-    // After deletions, check if the remaining points are still there.
-    let all_points = tree.range_search_bbox(&spart::geometry::Rectangle {
-        x: -1.0,
-        y: -1.0,
-        width: 12.0,
-        height: 12.0,
-    });
-    assert_eq!(all_points.len(), 7);
-
-    for i in 3..10 {
-        assert!(tree.delete(&points[i]));
-    }
-
-    let all_points_after_all_deleted = tree.range_search_bbox(&spart::geometry::Rectangle {
-        x: -1.0,
-        y: -1.0,
-        width: 12.0,
-        height: 12.0,
-    });
-    assert!(all_points_after_all_deleted.is_empty());
-}
-
-#[test]
-fn test_rstar_tree_empty() {
-    let mut tree: RStarTree<Point2D<&str>> = RStarTree::new(CAPACITY).unwrap();
-    let target = target_point_2d();
-
-    let knn_results = tree.knn_search::<EuclideanDistance>(&target, 5);
-    assert!(
-        knn_results.is_empty(),
-        "kNN search on empty tree should return no points"
-    );
-
-    let range_results = tree.range_search::<EuclideanDistance>(&target, 10.0);
-    assert!(
-        range_results.is_empty(),
-        "Range search on empty tree should return no points"
-    );
-
-    assert!(
-        !tree.delete(&target),
-        "Deleting from an empty tree should return false"
-    );
-}
-
-#[test]
-fn test_rstar_tree_knn_edge_cases() {
-    let mut tree: RStarTree<Point2D<&str>> = RStarTree::new(CAPACITY).unwrap();
-    let points = common_points_2d();
-    tree.insert_bulk(points.clone());
-
-    let target = target_point_2d();
-    let num_points = points.len();
-
-    let knn_results = tree.knn_search::<EuclideanDistance>(&target, 0);
-    assert!(
-        knn_results.is_empty(),
-        "kNN search with k=0 should return no points"
-    );
-
-    let knn_results = tree.knn_search::<EuclideanDistance>(&target, num_points + 5);
-    assert_eq!(
-        knn_results.len(),
-        num_points,
-        "kNN search with k > num_points should return all points"
-    );
-}
-
-#[test]
-fn test_rstar_tree_range_zero_radius() {
-    let mut tree: RStarTree<Point2D<&str>> = RStarTree::new(CAPACITY).unwrap();
-    let points = common_points_2d();
-    tree.insert_bulk(points.clone());
-
-    let target = points[0].clone();
-    let results = tree.range_search::<EuclideanDistance>(&target, 0.0);
-    assert_eq!(
-        results.len(),
-        1,
-        "Range search with zero radius should return only the exact point"
-    );
-    assert_eq!(*results[0], target);
-}
-
-#[test]
-fn test_rstar_tree_duplicates() {
-    let mut tree: RStarTree<Point2D<&str>> = RStarTree::new(CAPACITY).unwrap();
-    let p1 = common_points_2d()[0].clone();
-    let p2 = p1.clone();
-    tree.insert(p1.clone());
-    tree.insert(p2.clone());
-
-    let target = p1.clone();
-    let results = tree.knn_search::<EuclideanDistance>(&target, 2);
-    assert_eq!(results.len(), 2, "kNN should return duplicate points");
-
-    assert!(
-        tree.delete(&p1),
-        "Deleting a duplicate point should succeed"
-    );
-
-    let results_after_delete = tree.knn_search::<EuclideanDistance>(&target, 2);
-    assert_eq!(
-        results_after_delete.len(),
-        1,
-        "Deleting a point should only remove one instance"
-    );
-}
-
-#[test]
 fn test_rstar_tree_insert_bulk_3d() {
     let mut tree: RStarTree<Point3D<&str>> = RStarTree::new(CAPACITY).unwrap();
     let points = common_points_3d();
